# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Наша CRM",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/group/category": {
            "get": {
                "tags": [
                    "Группы"
                ],
                "summary": "Получить список категорий групп",
                "description": "Вернет список всех категорий групп текущего пользователя",
                "operationId": "078c843ef8e3fde32fd8291743dd5bc9",
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "categories": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": ""
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": ""
                                                    }
                                                },
                                                "type": "object"
                                            },
                                            "example": [
                                                {
                                                    "id": 1,
                                                    "name": "Fanger"
                                                },
                                                {
                                                    "id": 2,
                                                    "name": "Fanger"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/group/category/add": {
            "post": {
                "tags": [
                    "Группы"
                ],
                "summary": "Добавить категорию группы",
                "description": "Нужно передать поля name",
                "operationId": "2cae205d4eeee68c72178baa713ce8d1",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
        "/api/v1/group/category/update/{category_id}": {
            "put": {
                "tags": [
                    "Группы"
                ],
                "summary": "Обновить категорию группы",
                "description": "Нужно передать поля name",
                "operationId": "532738e53f3b8feca9256cbff4e6952b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "test"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "422": {
                        "description": "Вернет ошибку валидации, если поля не валидны или какие-то не отправлены"
                    }
                },
                "security": [
                    {
                        "bearer": []
                    }
                ]
            }
        },
...
    }
}